Iss104 - Capture fatal error during horizon config
During the quickstart installation, while horizon is being configured
via the postinst script, there is an opportunity for a DB migration to
be peformed.

In the event of this being the first run though, there is no stellar
role to execute the psql command as though and a FATAL error is thrown.

The script now tests for the existence of the stellar role and only
proceeds to DB migration if the role exists.